+++ plugin-parent/report/src/main/java/jscover/maven/MergeMojo.java
@@ -3,8 +3,6 @@ package jscover.maven;
 
 import jscover.report.ConfigurationForReport;
 import jscover.report.Main;
-import jscover.report.ReportFormat;
-import org.apache.maven.model.FileSet;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
@@ -13,23 +11,21 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
 import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 @Mojo(name = "merge", defaultPhase = LifecyclePhase.POST_INTEGRATION_TEST, threadSafe = true)
-public class MergetMojo extends AbstractMojo {
-    @Parameter
+public class MergeMojo extends AbstractMojo {
+    @Parameter(required = true)
     protected final List<String> mergeDirStrings = new ArrayList<String>();
-    @Parameter
-    private File destDir = new File("dest");
+    @Parameter(required = true)
+    private File destDir;
 
     public void execute() throws MojoExecutionException, MojoFailureException {
         Main main = new Main();
-        main.initialize();
         ConfigurationForReport config = getConfigurationForReport();
         main.setConfig(config);
+        main.initialize();
         main.mergeReports();
     }